Finish aa_4_2

+++ aa_4_2/func_matrix_mult.h
@@ -6,11 +6,11 @@
 
 typedef struct
 {
-	double **matrix_result;
-	double **matrix1;
-	double **matrix2;
-	double *row_factor;
-	double *col_factor;
+	int **matrix_result;
+	int **matrix1;
+	int **matrix2;
+	int *row_factor;
+	int *col_factor;
 	int i1;
 	int i2;
 	int col_count1;
@@ -18,12 +18,12 @@ typedef struct
 	int d;
 } function_resourse_t;
 
-error_t standard_multiply_matrix(double **matrix1, double **matrix2,
+error_t standard_multiply_matrix(int **matrix1, int **matrix2,
 		int row_count1, int col_count1, int row_count2, int col_count2,
-	       	thrd_t *threads, int threads_count, function_resourse_t *, double **matrix_result);
+	       	thrd_t *threads, int threads_count, function_resourse_t *, int **matrix_result);
 
-error_t vinograd_multiply_matrix(double **matrix1, double **matrix2, double *row_factor, double *col_factor,
+error_t vinograd_multiply_matrix(int **matrix1, int **matrix2, int *row_factor, int *col_factor,
 		int row_count1, int col_count1, int row_count2, int col_count2,
-	       	thrd_t *threads, int threads_count, function_resourse_t *, double **matrix_result);
+	       	thrd_t *threads, int threads_count, function_resourse_t *, int **matrix_result);
 
 #endif
